+++ dew/src/main/java/io/github/ikws4/dew/core/gl/resource/Time.java
@@ -1,5 +1,5 @@
 package io.github.ikws4.dew.core.gl.resource;
 
 public class Time {
-  public double deltaTime;
+  public float deltaTime;
 }
